+++ CHANGES.md
@@ -7,6 +7,7 @@ but this file may sometimes contain later improvements (e.g. typo fixes).
 
 ## next (not yet released)
 
+- Improved error handling in `completeOAuthSession()`.
 - Updated dependencies.
 
 ## v0.3.1 (2024-09-07)